pyproject.toml
[project]
name = "persona-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
]
